requirements.txt
PyMuPDF>=1.24.0  
python-pptx>=1.0.2
nltk>=3.9.1
openai==1.82.0
tenacity>=8.2.3
neo4j>=5.14.0           
tiktoken==0.5.1                
python-dotenv==1.0.1 
rich==13.7.1
boto3==1.38.31
azure-storage-blob==12.25.1
neo4j==5.28.1
google==3.0.0
google-cloud-storage==3.1.0
fastapi==0.115.12